+++ bin/workflow_glue/check_sample_sheet_condition.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 """Check if a sample sheet is valid."""
+from collections import Counter
 import csv
 import sys
 
@@ -11,30 +12,27 @@ def main(args):
     logger = get_named_logger("checkSheetCondition")
     with open(args.sample_sheet, "r") as f:
         csv_reader = csv.DictReader(f)
-        unique_controls = []
-        controls_dic = {}
+        conditions_count = Counter()
         for row in csv_reader:
-            if 'condition' in list(row.keys()):
-                unique_controls.append(row['condition'])
-                if row['condition'] not in controls_dic:
-                    controls_dic[row['condition']] = 1
-                else:
-                    controls_dic[row['condition']] += 1
+            if "condition" in row:
+                conditions_count[row['condition']] += 1
             else:
                 sys.exit(
                     "Sample sheet has no condition column "
                     "which is required for the "
                     "differential expression subworkflow.")
-        if len(list(set(controls_dic.keys()))) != 2:
+        if len(conditions_count.keys()) != 2:
             sys.exit(
                 "There must be only two unique conditions "
                 "in the condition column of the sample sheet.")
-        for val in list(controls_dic.values()):
-            if val < 2:
-                sys.exit(
-                    "There must be at least 2 repeats for each "
-                    "condition indicated in the sample sheet.")
-
+        if "control" not in conditions_count:
+            sys.exit(
+                "One of the condition types must be control, "
+                "to indicate which samples to use as the reference.")
+        if any(v < 2 for v in conditions_count.values()):
+            sys.exit(
+                "There must be at least 2 repeats for each "
+                "condition indicated in the sample sheet.")
     logger.info(f"Checked sample sheet for condition column {args.sample_sheet}.")