+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+- Sample sheet must include a `control` type to indicate which samples are the reference for the differential expression pipeline.
+
 ## [v0.4.1]
 ### Changed
 - Updated docker container with Pychopper to support LSK114.